+++ src/Text/Show/Text/Text/Read.hs
@@ -18,40 +18,36 @@ module Text.Show.Text.Text.Read (
 #endif
     ) where
 
-import           Data.Text.Lazy.Builder (Builder, fromString)
+import Data.Text.Lazy.Builder (Builder)
 
-import           GHC.Show (appPrec, appPrec1)
+import Prelude hiding (Show)
 
-import           Prelude hiding (Show)
-
-import           Text.Read.Lex (Lexeme(..))
+import Text.Read.Lex (Lexeme(..))
 #if MIN_VERSION_base(4,7,0)
-import           Text.Read.Lex (Number)
+import Text.Read.Lex (Number)
 #endif
-import           Text.Show.Text.Classes (Show(showbPrec), showbParen)
-import           Text.Show.Text.Data.Char (showbLitChar)
+import Text.Show.Text.Classes (Show(showbPrec), showbUnary)
 #if MIN_VERSION_base(4,6,0)
-import qualified Text.Show as S (showsPrec)
+import Text.Show.Text.Newtypes (FromStringShow(..), LitChar(..), LitString(..))
 #else
-import           Text.Show.Text.Data.Integral (showbIntegerPrec)
-import           Text.Show.Text.Data.Ratio (showbRatioPrec)
+import Text.Show.Text.Data.Integral (showbIntegerPrec)
+import Text.Show.Text.Data.Ratio (showbRatioPrec)
 #endif
-import           Text.Show.Text.Utils ((<>))
 
 #include "inline.h"
 
 -- | Convert a 'Lexeme' to a 'Builder' with the given precedence.
 showbLexemePrec :: Int -> Lexeme -> Builder
-showbLexemePrec p (Char c)   = showbParen (p > appPrec) $ "Char "   <> showbLitChar c
-showbLexemePrec p (String s) = showbParen (p > appPrec) $ "String " <> fromString s
-showbLexemePrec p (Punc pun) = showbParen (p > appPrec) $ "Punc "   <> fromString pun
-showbLexemePrec p (Ident i)  = showbParen (p > appPrec) $ "Ident "  <> fromString i
-showbLexemePrec p (Symbol s) = showbParen (p > appPrec) $ "Symbol " <> fromString s
+showbLexemePrec p (Char c)   = showbUnary "Char"   p $ LitChar c
+showbLexemePrec p (String s) = showbUnary "String" p $ LitString s
+showbLexemePrec p (Punc pun) = showbUnary "Punc"   p $ LitString pun
+showbLexemePrec p (Ident i)  = showbUnary "Ident"  p $ LitString i
+showbLexemePrec p (Symbol s) = showbUnary "Symbol" p $ LitString s
 #if MIN_VERSION_base(4,6,0)
-showbLexemePrec p (Number n) = showbParen (p > appPrec) $ "Number " <> fromString (S.showsPrec appPrec1 n "")
+showbLexemePrec p (Number n) = showbUnary "Number" p $ FromStringShow n
 #else
-showbLexemePrec p (Int i)    = showbParen (p > appPrec) $ "Int "    <> showbIntegerPrec appPrec1 i
-showbLexemePrec p (Rat r)    = showbParen (p > appPrec) $ "Rat "    <> showbRatioPrec appPrec1 r
+showbLexemePrec p (Int i)    = showbUnary "Int" p i
+showbLexemePrec p (Rat r)    = showbUnary "Rat" p r
 #endif
 showbLexemePrec _ EOF        = "EOF"
 {-# INLINE showbLexemePrec #-}
@@ -59,7 +55,7 @@ showbLexemePrec _ EOF        = "EOF"
 #if MIN_VERSION_base(4,7,0)
 -- | Convert a 'Number' to a 'Builder' with the given precedence.
 showbNumberPrec :: Int -> Number -> Builder
-showbNumberPrec p n = fromString $ S.showsPrec p n ""
+showbNumberPrec p = showbPrec p . FromStringShow
 {-# INLINE showbNumberPrec #-}
 #endif